Split types, cleanup
