+++ src/shared/types/user.ts
@@ -1,0 +1,8 @@
+export type WorkStatus = "looking" | "passive" | "not_looking";
+
+export interface UserProfile {
+  id: string;
+  name: string;
+  avatar: string;
+  workStatus: WorkStatus;
+}
